workflow/service_chaining/train_model/requirements.txt
joblib
pandas
scikit-learn
flask>=1.0.2
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
